requirements.txt
streamlit
sentence-transformers
transformers
torch
faiss-cpu
pdfplumber
nltk
